fix PageFault from kernel and deadlock
1. when calling GetNextForCpu outside VM, it seems the cache is not flushed
in a timely manner and sometimes it will return a running task, that
will make two vcpu running a same task simultanuously, which will be a
disaster. we hold the queue lock while check the fields in the task, to
make sure the cache is flushed while we read them.

2. fix the deadlock in the task_signals.rs